// src/generated/protobuf/shared.ts
export enum Kind {
  AUDIO = 0,
  VIDEO = 1,
  UNRECOGNIZED = -1,
}
// ...
export enum BitrateControlMode {
  DYNAMIC_CONSUMERS = 0,
  MAX_BITRATE = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface Sender {
  kind: Kind;
  name: string;
  state: Sender_State | undefined;
}
// ...
export interface Sender_Source {
  id: string;
  screen: boolean;
  metadata?: string | undefined;
}

export interface Sender_Config {
  priority: number;
  bitrate: BitrateControlMode;
}

export interface Sender_State {
  config: Sender_Config | undefined;
  source?: Sender_Source | undefined;
}
// ...
export const Sender = {
  encode(
    message: Sender,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.kind !== 0) {
      writer.uint32(8).int32(message.kind);
    }
    if (message.name !== '') {
      writer.uint32(18).string(message.name);
    }
    if (message.state !== undefined) {
      Sender_State.encode(message.state, writer.uint32(26).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Sender {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSender();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.kind = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.state = Sender_State.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Sender {
    return {
      kind: isSet(object.kind) ? kindFromJSON(object.kind) : 0,
      name: isSet(object.name) ? globalThis.String(object.name) : '',
      state: isSet(object.state)
        ? Sender_State.fromJSON(object.state)
        : undefined,
    };
  },

  toJSON(message: Sender): unknown {
    const obj: any = {};
    if (message.kind !== 0) {
      obj.kind = kindToJSON(message.kind);
    }
    if (message.name !== '') {
      obj.name = message.name;
    }
    if (message.state !== undefined) {
      obj.state = Sender_State.toJSON(message.state);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Sender>, I>>(base?: I): Sender {
    return Sender.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Sender>, I>>(object: I): Sender {
    const message = createBaseSender();
    message.kind = object.kind ?? 0;
    message.name = object.name ?? '';
    message.state =
      object.state !== undefined && object.state !== null
        ? Sender_State.fromPartial(object.state)
        : undefined;
    return message;
  },
};

function createBaseSender_Source(): Sender_Source {
  return { id: '', screen: false, metadata: undefined };
}

export const Sender_Source = {
  encode(
    message: Sender_Source,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.id !== '') {
      writer.uint32(10).string(message.id);
    }
    if (message.screen !== false) {
      writer.uint32(16).bool(message.screen);
    }
    if (message.metadata !== undefined) {
      writer.uint32(26).string(message.metadata);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Sender_Source {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSender_Source();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.screen = reader.bool();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.metadata = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Sender_Source {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : '',
      screen: isSet(object.screen) ? globalThis.Boolean(object.screen) : false,
      metadata: isSet(object.metadata)
        ? globalThis.String(object.metadata)
        : undefined,
    };
  },

  toJSON(message: Sender_Source): unknown {
    const obj: any = {};
    if (message.id !== '') {
      obj.id = message.id;
    }
    if (message.screen !== false) {
      obj.screen = message.screen;
    }
    if (message.metadata !== undefined) {
      obj.metadata = message.metadata;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Sender_Source>, I>>(
    base?: I,
  ): Sender_Source {
    return Sender_Source.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Sender_Source>, I>>(
    object: I,
  ): Sender_Source {
    const message = createBaseSender_Source();
    message.id = object.id ?? '';
    message.screen = object.screen ?? false;
    message.metadata = object.metadata ?? undefined;
    return message;
  },
};

function createBaseSender_Config(): Sender_Config {
  return { priority: 0, bitrate: 0 };
}

export const Sender_Config = {
  encode(
    message: Sender_Config,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.priority !== 0) {
      writer.uint32(8).uint32(message.priority);
    }
    if (message.bitrate !== 0) {
      writer.uint32(16).int32(message.bitrate);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Sender_Config {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSender_Config();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.priority = reader.uint32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.bitrate = reader.int32() as any;
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Sender_Config {
    return {
      priority: isSet(object.priority) ? globalThis.Number(object.priority) : 0,
      bitrate: isSet(object.bitrate)
        ? bitrateControlModeFromJSON(object.bitrate)
        : 0,
    };
  },

  toJSON(message: Sender_Config): unknown {
    const obj: any = {};
    if (message.priority !== 0) {
      obj.priority = Math.round(message.priority);
    }
    if (message.bitrate !== 0) {
      obj.bitrate = bitrateControlModeToJSON(message.bitrate);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Sender_Config>, I>>(
    base?: I,
  ): Sender_Config {
    return Sender_Config.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Sender_Config>, I>>(
    object: I,
  ): Sender_Config {
    const message = createBaseSender_Config();
    message.priority = object.priority ?? 0;
    message.bitrate = object.bitrate ?? 0;
    return message;
  },
};

function createBaseSender_State(): Sender_State {
  return { config: undefined, source: undefined };
}

export const Sender_State = {
  encode(
    message: Sender_State,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.config !== undefined) {
      Sender_Config.encode(message.config, writer.uint32(10).fork()).ldelim();
    }
    if (message.source !== undefined) {
      Sender_Source.encode(message.source, writer.uint32(18).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Sender_State {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSender_State();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.config = Sender_Config.decode(reader, reader.uint32());
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.source = Sender_Source.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Sender_State {
    return {
      config: isSet(object.config)
        ? Sender_Config.fromJSON(object.config)
        : undefined,
      source: isSet(object.source)
        ? Sender_Source.fromJSON(object.source)
        : undefined,
    };
  },

  toJSON(message: Sender_State): unknown {
    const obj: any = {};
    if (message.config !== undefined) {
      obj.config = Sender_Config.toJSON(message.config);
    }
    if (message.source !== undefined) {
      obj.source = Sender_Source.toJSON(message.source);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Sender_State>, I>>(
    base?: I,
  ): Sender_State {
    return Sender_State.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Sender_State>, I>>(
    object: I,
  ): Sender_State {
    const message = createBaseSender_State();
    message.config =
      object.config !== undefined && object.config !== null
        ? Sender_Config.fromPartial(object.config)
        : undefined;
    message.source =
      object.source !== undefined && object.source !== null
        ? Sender_Source.fromPartial(object.source)
        : undefined;
    return message;
  },
};
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };